Move payment processor list endpoint to api/
XCOM-211
